streaming/requirements.txt
# Streaming Pipeline — Week 1 dependencies
# Shreyas Kaldate (sk12898)
# Install: pip install -r requirements.txt

kafka-python==2.0.2
requests==2.31.0
